// apps/game-client/src/lib/api/generated/main/model/role-response-dto-output.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { RoleResponseDtoOutputPermissionsItem } from './role-response-dto-output-permissions-item';

export interface RoleResponseDtoOutput {
  id: string;
  guildId: string;
  name: string;
  /** @nullable */
  color: number | null;
  /** @nullable */
  position?: number | null;
  permissions: RoleResponseDtoOutputPermissionsItem[];
  /** @nullable */
  lvlRangeFrom?: number | null;
  /** @nullable */
  lvlRangeTo?: number | null;
}